backend/foodgram/requirements.txt
Django==3.2.3
Pillow==9.0.0
djangorestframework==3.12.4
django-filter==23.1
djoser==2.1.0
environs==11.2.0
gunicorn==20.1.0
jsonschema==4.23.0
psycopg2-binary==2.9.3
python-dotenv==1.0.1
flake8==6.0.0
flake8-isort==6.0.0
shortuuid==1.0.13
